+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - Allows for unrestricted SMBIOS usage with `AppleIntelSNBGraphicsFB`
   - Adds OpenCL downgrade for TeraScale 2
     - Resolves VNC support (credit IronApple#2711)
+  - Fix SecureBootModel detection
 - Add `OpenCore-Legacy-Patcher.plist` for applied patch info
   - Located under `/System/Library/CoreServices`
   - Lists patch sets applied including files installed and removed